Eliminate .cmp files
This is a conservative approach. The bytes that would have been written
to .cmp files, as a side effect of preprocessing, are now saved in a
string in each preprocessed module instead. The bytes are then dumped
into the .out files when the binary runs.

The reporter loads the bytes from the .out files instead of looking for
.cmp files. The data is deduplicated, since multiple .out files may
contain .cmp data for the same source files.

This means that the .cmp data is marshalled (as before) into a string,
but now that string is again marshalled when written out to the .out
file. That is not the right approach to encoding .cmp data in the long
term, but I am saving changing the representation for future commits,
in order to minimize this one.

ocveralls depends on certain behavior of
Bisect.Common.read_runtime_data and Bisect.Common.read_points. As a
temporary measure, the new versions of these functions are
distinguished with a ', and the old functions are implemented in terms
of the new. The implementation is fragile, however, and this interface
will have to be replaced.